@@ -1,5 +1,6 @@
-::Admin.controllers :base do
-  get :index, :map => "/" do
-    render "base/index"
+# frozen_string_literal: true
+TokyoProject::Admin.controllers :base do
+  get :index, map: '/' do
+    render 'base/index'
   end
 end
