@@ -1,7 +1,7 @@
-Admin.controllers :pictures do
-
+# frozen_string_literal: true
+TokyoProject::Admin.controllers :pictures do
   get :index do
-    @pictures = Picture.order('created_at DESC').paginate(:page => params[:page], :per_page => 10)
+    @pictures = Picture.order('created_at DESC').page(params[:page]).per(10)
     render 'pictures/index'
   end
 
@@ -15,7 +15,7 @@ Admin.controllers :pictures do
     @picture = Picture.new(params[:picture])
     if @picture.save
       flash[:notice] = 'Picture was successfully created.'
-      redirect url(:pictures, :edit, :id => @picture.id)
+      redirect url(:pictures, :edit, id: @picture.id)
     else
       flash[:error] = 'Picture is not valid.'
       redirect url(:pictures, :new)
@@ -23,13 +23,13 @@ Admin.controllers :pictures do
     @picture.clear_cache
   end
 
-  get :edit, :with => :id do
+  get :edit, with: :id do
     @picture = Picture.find(params[:id])
     @visions = Vision.all
     render 'pictures/edit'
   end
 
-  put :update, :with => :id do
+  put :update, with: :id do
     @picture = Picture.find(params[:id])
     if @picture.update_attributes(params[:picture])
       @picture.reload
@@ -38,13 +38,13 @@ Admin.controllers :pictures do
         PictureJob.new.async.schedule(@picture.seconds_to_publish, @picture.id)
       end
       flash[:notice] = 'Picture was successfully updated.'
-      redirect url(:pictures, :edit, :id => @picture.id)
+      redirect url(:pictures, :edit, id: @picture.id)
     else
       render 'pictures/edit'
     end
   end
 
-  delete :destroy, :with => :id do
+  delete :destroy, with: :id do
     picture = Picture.find(params[:id])
     vision = picture.vision
     if picture.destroy
