@@ -1,6 +1,7 @@
-::Admin.controllers :sessions do
+# frozen_string_literal: true
+TokyoProject::Admin.controllers :sessions do
   get :new do
-    render "/sessions/new", nil, :layout => false
+    render '/sessions/new', nil, layout: false
   end
 
   post :create do
@@ -12,9 +13,9 @@
       set_current_account(account)
       redirect url(:base, :index)
     else
-      params[:email], params[:password] = h(params[:email]), h(params[:password])
-      flash[:error] = pat('login.error')
-      redirect url(:sessions, :new)
+      params[:email] = h(params[:email])
+      flash.now[:error] = pat('login.error')
+      render '/sessions/new', nil, layout: false
     end
   end
 
