@@ -1,7 +1,7 @@
-Admin.controllers :visions do
-
+# frozen_string_literal: true
+TokyoProject::Admin.controllers :visions do
   get :index do
-    @visions = Vision.order('created_at DESC').paginate(:page => params[:page], :per_page => 10)
+    @visions = Vision.order('created_at DESC').page(params[:page]).per(10)
     render 'visions/index'
   end
 
@@ -16,29 +16,29 @@ Admin.controllers :visions do
     @vision.account = current_account
     if @vision.save
       flash[:notice] = 'Vision was successfully created.'
-      redirect url(:visions, :edit, :id => @vision.id)
+      redirect url(:visions, :edit, id: @vision.id)
     else
       render 'visions/new'
     end
   end
 
-  get :edit, :with => :id do
+  get :edit, with: :id do
     @vision = Vision.find(params[:id])
     @areas = Area.all
     render 'visions/edit'
   end
 
-  put :update, :with => :id do
+  put :update, with: :id do
     @vision = Vision.find(params[:id])
     if @vision.update_attributes(params[:vision])
       flash[:notice] = 'Vision was successfully updated.'
-      redirect url(:visions, :edit, :id => @vision.id)
+      redirect url(:visions, :edit, id: @vision.id)
     else
       render 'visions/edit'
     end
   end
 
-  delete :destroy, :with => :id do
+  delete :destroy, with: :id do
     vision = Vision.find(params[:id])
     if vision.destroy
       flash[:notice] = 'Vision was successfully destroyed.'
