@@ -1,6 +1,7 @@
-TokyoProject.controllers :atom do
-  get :index, :provides => [:atom] do
-    @pictures = Picture.all
+# frozen_string_literal: true
+TokyoProject::App.controllers :atom do
+  get :index, provides: [:atom] do
+    @pictures = Picture.published.last(20)
     render 'atom/index'
   end
 end
