@@ -1,9 +1,10 @@
-TokyoProject::TokyoProject.controllers :maps do
-
-  get :show, :map => '/maps/:id' do
+# frozen_string_literal: true
+TokyoProject::App.controllers :maps do
+  get :show, map: '/maps/:id' do
     key = "maps_show_#{params[:id]}"
+    cache_time = Padrino.env == :production ? 86_400 : 1
 
-    cache(key, expires_in: (Padrino.env.to_s == "production" ? 3600 : 1)) do
+    cache(key, expires: cache_time) do
       @vision = Vision.find(params[:id])
       map_data = @vision.map_data
       @static_map_url = map_data['static_url']
@@ -12,5 +13,4 @@ TokyoProject::TokyoProject.controllers :maps do
       render 'maps/show'
     end
   end
-
 end
