@@ -1,52 +1,68 @@
+# frozen_string_literal: true
+
 cleardb = URI.parse(ENV['CLEARDB_DATABASE_URL'] || '')
+mariadb = URI.parse(ENV['JAWSDB_MARIA_URL'] || '')
 
 ActiveRecord::Base.configurations[:development] = {
-  :adapter   => 'mysql2',
-  :encoding  => 'utf8',
-  :reconnect => true,
-  :database  => 'tokyo_project_development',
-  :pool      => 5,
-  :username  => 'root',
-  :password  => '',
-  :host      => 'localhost',
-  :socket    => '/tmp/mysql.sock'
+  adapter: 'mysql2',
+  encoding: 'utf8',
+  reconnect: true,
+  database: 'tokyo_project_development',
+  pool: 5,
+  username: 'root',
+  password: '',
+  host: 'localhost',
+  socket: '/tmp/mysql.sock'
 
 }
 
+# ActiveRecord::Base.configurations[:production] = {
+#   adapter: 'mysql2',
+#   encoding: 'utf8',
+#   reconnect: true,
+#   database: cleardb.path[1..-1],
+#   username: cleardb.user,
+#   password: cleardb.password,
+#   host: cleardb.host
+# }
+
 ActiveRecord::Base.configurations[:production] = {
-  :adapter   => 'mysql2',
-  :encoding  => 'utf8',
-  :reconnect => true,
-  :database  => cleardb.path[1..-1],
-  :username  => cleardb.user,
-  :password  => cleardb.password,
-  :host      => cleardb.host
+  adapter: 'mysql2',
+  encoding: 'utf8',
+  reconnect: true,
+  database: mariadb.path[1..-1],
+  username: mariadb.user,
+  password: mariadb.password,
+  host: mariadb.host
 }
 
 ActiveRecord::Base.configurations[:test] = {
-  :adapter   => 'mysql2',
-  :encoding  => 'utf8',
-  :reconnect => true,
-  :database  => 'tokyo_project_test',
-  :pool      => 5,
-  :username  => 'root',
-  :password  => '',
-  :host      => 'localhost',
-  :socket    => '/tmp/mysql.sock'
+  adapter: 'mysql2',
+  encoding: 'utf8',
+  reconnect: true,
+  database: 'tokyo_project_test',
+  pool: 5,
+  username: 'root',
+  password: '',
+  host: 'localhost',
+  socket: '/tmp/mysql.sock'
 
 }
+#
 
 # Setup our logger
 ActiveRecord::Base.logger = logger
 
-# Raise exception on mass assignment protection for Active Record models
-ActiveRecord::Base.mass_assignment_sanitizer = :strict
+if ActiveRecord::VERSION::MAJOR.to_i < 4
+  # Raise exception on mass assignment protection for Active Record models.
+  ActiveRecord::Base.mass_assignment_sanitizer = :strict
 
-# Log the query plan for queries taking more than this (works
-# with SQLite, MySQL, and PostgreSQL)
-ActiveRecord::Base.auto_explain_threshold_in_seconds = 0.5
+  # Log the query plan for queries taking more than this (works
+  # with SQLite, MySQL, and PostgreSQL).
+  ActiveRecord::Base.auto_explain_threshold_in_seconds = 0.5
+end
 
-# Include Active Record class name as root for JSON serialized output.
+# Doesn't include Active Record class name as root for JSON serialized output.
 ActiveRecord::Base.include_root_in_json = false
 
 # Store the full class name (including module namespace) in STI type column.
@@ -55,9 +71,12 @@ ActiveRecord::Base.store_full_sti_class = true
 # Use ISO 8601 format for JSON serialized times and dates.
 ActiveSupport.use_standard_json_time_format = true
 
-# Don't escape HTML entities in JSON, leave that for the #json_escape helper.
-# if you're including raw json in an HTML page.
+# Don't escape HTML entities in JSON, leave that for the #json_escape helper
+# if you're including raw JSON in an HTML page.
 ActiveSupport.escape_html_entities_in_json = false
 
-# Now we can estabilish connection with our db
+# Now we can establish connection with our db.
 ActiveRecord::Base.establish_connection(ActiveRecord::Base.configurations[Padrino.env])
+
+# Timestamps are in the utc by default.
+ActiveRecord::Base.default_timezone = :utc
