@@ -1,31 +1,30 @@
+# frozen_string_literal: true
 # Seed add you the ability to populate your db.
 # We provide you a basic shell for interaction with the end user.
 # So try some code like below:
 #
-name = shell.ask("What's your name?")
-shell.say name
+#   name = shell.ask("What's your name?")
+#   shell.say name
+#
+email     = shell.ask 'Which email do you want use for logging into admin?'
+password  = shell.ask 'Tell me the password to use:', echo: false
 
-email     = shell.ask "Which email do you want use for logging into admin?"
-password  = shell.ask "Tell me the password to use:"
+shell.say ''
 
-shell.say ""
-
-account = Account.create(:email => email, :name => "Foo", :surname => "Bar", :password => password, :password_confirmation => password, :role => "admin")
+account = Account.new(email: email, name: 'Foo', surname: 'Bar', password: password, password_confirmation: password, role: 'admin')
 
 if account.valid?
-  shell.say "================================================================="
-  shell.say "Account has been successfully created, now you can login with:"
-  shell.say "================================================================="
+  account.save
+  shell.say '================================================================='
+  shell.say 'Account has been successfully created, now you can login with:'
+  shell.say '================================================================='
   shell.say "   email: #{email}"
-  shell.say "   password: #{password}"
-  shell.say "================================================================="
+  shell.say "   password: #{'*' * password.length}"
+  shell.say '================================================================='
 else
-  shell.say "Sorry but some thing went wrong!"
-  shell.say ""
+  shell.say 'Sorry, but something went wrong!'
+  shell.say ''
   account.errors.full_messages.each { |m| shell.say "   - #{m}" }
 end
 
-shell.say "Generating Wards"
-wards_names = %w[Adachi Arakawa Bunkyō Chiyoda Chūō Edogawa Itabashi Katsushika Kita Kōtō Meguro Minato Nakano Nerima Ōta Setagaya Shibuya Shinagawa Shinjuku Suginami Sumida Taitō Toshima]
-wards_names.each{|w| Ward.create(:name => w)}
-shell.say "Wards Created"
+shell.say ''
