@@ -1,28 +1,26 @@
+# frozen_string_literal: true
 class Account < ActiveRecord::Base
   attr_accessor :password, :password_confirmation
 
   # Validations
   validates_presence_of     :email, :role
-  validates_presence_of     :password,                   :if => :password_required
-  validates_presence_of     :password_confirmation,      :if => :password_required
-  validates_length_of       :password, :within => 4..40, :if => :password_required
-  validates_confirmation_of :password,                   :if => :password_required
-  validates_length_of       :email,    :within => 3..100
-  validates_uniqueness_of   :email,    :case_sensitive => false
-  validates_format_of       :email,    :with => /\A([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})\Z/i
-  validates_format_of       :role,     :with => /[A-Za-z]/
+  validates_presence_of     :password,                   if: :password_required
+  validates_presence_of     :password_confirmation,      if: :password_required
+  validates_length_of       :password, within: 4..40, if: :password_required
+  validates_confirmation_of :password, if: :password_required
+  validates_length_of       :email,    within: 3..100
+  validates_uniqueness_of   :email,    case_sensitive: false
+  validates_format_of       :email,    with: /\A([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})\Z/i
+  validates_format_of       :role,     with: /[A-Za-z]/
 
   # Callbacks
-  before_save :encrypt_password, :if => :password_required
-
-  has_many :visions
-  has_many :posts
+  before_save :encrypt_password, if: :password_required
 
   ##
-  # This method is for authentication purpose
+  # This method is for authentication purpose.
   #
   def self.authenticate(email, password)
-    account = first(:conditions => { :email => email }) if email.present?
+    account = where('lower(email) = lower(?)', email).first if email.present?
     account && account.has_password?(password) ? account : nil
   end
 
@@ -31,8 +29,11 @@ class Account < ActiveRecord::Base
   end
 
   private
+
   def encrypt_password
-    self.crypted_password = ::BCrypt::Password.create(password)
+    value = ::BCrypt::Password.create(password)
+    value = value.force_encoding(Encoding::UTF_8) if value.encoding == Encoding::ASCII_8BIT
+    self.crypted_password = value
   end
 
   def password_required
