@@ -1,54 +1,61 @@
+# frozen_string_literal: true
 class Picture < ActiveRecord::Base
   belongs_to :vision
-  scope :published, where(:is_published => true)
   validates_uniqueness_of :flickr_id
 
+  scope :published, -> { where(is_published: true) }
+
   def flickr_version
-    FlickrConnector.new.get_picture(self.flickr_id)
+    FlickrConnector.new.get_picture(flickr_id)
   end
 
   def thumb
-    url = TokyoProject.cache.get "cached_picture_#{self.id}_tumb"
-    if url == "" || url == nil
-      url = FlickRaw.url_t(flickr_version) rescue "/images/default_small.jpg"
-      TokyoProject.cache.set("cached_picture_#{self.id}_tumb", url, :expires_in => 86400)
+    cache_key = "cached_picture_#{id}_tumb"
+    Padrino.cache.fetch(cache_key, expires: 86_400) do
+      FlickRaw.url_t(flickr_version)
     end
-    return url
   end
 
   def medium
-    url = TokyoProject.cache.get "cached_picture_#{self.id}_medium"
-    if url == "" || url == nil
-      url = FlickRaw.url_z(flickr_version) rescue ""
-      TokyoProject.cache.set("cached_picture_#{self.id}_medium", url, :expires_in => 86400)
+    cache_key = "cached_picture_#{id}_medium"
+    Padrino.cache.fetch(cache_key, expires: 86_400) do
+      FlickRaw.url_z(flickr_version)
     end
-    return url
   end
 
   def large
-    url = TokyoProject.cache.get "cached_picture_#{self.id}_large"
-    if url == "" || url == nil
-      url = FlickRaw.url_b(flickr_version) rescue ""
-      TokyoProject.cache.set("cached_picture_#{self.id}_large", url, :expires_in => 86400)
+    cache_key = "cached_picture_#{id}_large"
+    Padrino.cache.fetch(cache_key, expires: 86_400) do
+      FlickRaw.url_b(flickr_version)
     end
-    return url
   end
 
-  def pinpoint_code(label="1")
-    code = (self.latitude && self.longitude) ? "&markers=|color:blue|label:#{label}|#{self.latitude},#{self.longitude}" : ""
+  def pinpoint_code(label = '1')
+    code = nil
+    if latitude && longitude
+      code = "&markers=|color:blue|label:#{label}|#{latitude},#{longitude}"
+    else
+      code ''
+    end
+    code
   end
 
   def clear_cache
-    #should clear cache for itself, the vision and the area
-    key = "cached_picture_#{self.id}"
-    ['_tumb', '_medium', '_large'].each{|suffix| TokyoProject.cache.delete(key+suffix)}
+    # should clear cache for itself, the vision and the area
+    key = "cached_picture_#{id}"
 
-    keys = ["vision_show_#{self.vision.id}", "vision_show_#{self.vision.id}_#{self.id}"]
-    keys.each{|k| TokyoProject.cache.delete(k)}
+    %w(_tumb _medium _large).each do |suffix|
+      Padrino.cache.delete(key + suffix)
+    end
+
+    keys = ["vision_show_#{vision.id}", "vision_show_#{vision.id}_#{id}"]
+    keys.each { |k| Padrino.cache.delete(k) }
     vision.clear_cache
   end
 
   def seconds_to_publish
-    schedule_at.to_i - Time.now.to_i rescue 0
+    schedule_at.to_i - Time.now.to_i
+  rescue
+    0
   end
 end
