@@ -1,11 +1,14 @@
+# frozen_string_literal: true
 require 'spec_helper'
 
-describe "MapsController" do
-  before do
-    get "/"
-  end
+RSpec.describe '/maps' do
+  pending "add some examples to #{__FILE__}" do
+    before do
+      get '/maps'
+    end
 
-  it "returns hello world" do
-    last_response.body.should == "Hello World"
+    it 'returns hello world' do
+      expect(last_response.body).to eq 'Hello World'
+    end
   end
 end
