@@ -1,7 +1,14 @@
+# frozen_string_literal: true
 # Helper methods defined here can be accessed in any controller or view in the application
 
-TokyoProject.helpers do
-  # def simple_helper_method
-  #  ...
-  # end
+module TokyoProject
+  class App
+    module RssHelper
+      # def simple_helper_method
+      # ...
+      # end
+    end
+
+    helpers RssHelper
+  end
 end
