@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 class Area < ActiveRecord::Base
   belongs_to :ward
   has_many :visions
@@ -5,13 +6,22 @@ class Area < ActiveRecord::Base
   before_save :update_url_title
 
   def clear_cache
-    TokyoProject.cache.delete("root_path")
-    TokyoProject.cache.delete("areas")
-    TokyoProject.cache.delete("area_show_#{self.id}")
+    Padrino.cache.delete('root_path')
+    Padrino.cache.delete('areas')
+    Padrino.cache.delete("area_show_#{id}")
   end
 
   def update_url_title
-    tmp_title = self.name.gsub(',', '').gsub('.', '').split(' ').collect{|word| word.capitalize}.join('').underscore rescue nil
-    self.url_title= URI::encode(tmp_title)
+    tmp_title = begin
+                  name.delete(',')
+                      .delete('.')
+                      .split(' ')
+                      .collect(&:capitalize)
+                      .join('')
+                      .underscore
+                rescue
+                  nil
+                end
+    self.url_title = URI.encode(tmp_title)
   end
 end
