@@ -1,10 +1,8 @@
+# frozen_string_literal: true
 class Post < ActiveRecord::Base
-  belongs_to :account
-
   def clear_cache
-    #should clear cache for itself, and the index
-    keys = ["blog_show_#{self.id}", "blog"]
-    keys.each{|k| TokyoProject.cache.delete(k)}
+    # should clear cache for itself, and the index
+    keys = ["blog_show_#{id}", 'blog']
+    keys.each { |k| Padrino.cache.delete(k) }
   end
-
 end
