@@ -112,6 +112,7 @@
     }
 
     // Autofocus first field with an error. (usability)
-    $('.has-error :input').first().focus();
+    var error_input;
+    if (error_input = $('.has-error :input').first()) { error_input.focus(); }
   });
 }(window.jQuery);
