@@ -1,8 +1,6 @@
+# frozen_string_literal: true
 require 'spec_helper'
 
-describe "Account Model" do
-  let(:account) { Account.new }
-  it 'can be created' do
-    account.should_not be_nil
-  end
+RSpec.describe Account do
+  pending "add some examples to (or delete) #{__FILE__}"
 end
