@@ -1,8 +1,6 @@
+# frozen_string_literal: true
 require 'spec_helper'
 
-describe "Area Model" do
-  let(:area) { Area.new }
-  it 'can be created' do
-    area.should_not be_nil
-  end
+RSpec.describe Area do
+  pending "add some examples to (or delete) #{__FILE__}"
 end
