@@ -1,4 +1,6 @@
+# frozen_string_literal: true
 require 'spec_helper'
 
-describe Post do
+RSpec.describe Post do
+  pending "add some examples to (or delete) #{__FILE__}"
 end
