@@ -1,8 +1,6 @@
+# frozen_string_literal: true
 require 'spec_helper'
 
-describe "Ward Model" do
-  let(:ward) { Ward.new }
-  it 'can be created' do
-    ward.should_not be_nil
-  end
+RSpec.describe Ward do
+  pending "add some examples to (or delete) #{__FILE__}"
 end
