@@ -1,13 +1,22 @@
-PADRINO_ENV = 'test' unless defined?(PADRINO_ENV)
-require File.expand_path(File.dirname(__FILE__) + "/../config/boot")
+# frozen_string_literal: true
+RACK_ENV = 'test' unless defined?(RACK_ENV)
+require File.expand_path(File.dirname(__FILE__) + '/../config/boot')
+Dir[File.expand_path(File.dirname(__FILE__) + '/../app/helpers/**/*.rb')].each(&method(:require))
 
 RSpec.configure do |conf|
   conf.include Rack::Test::Methods
 end
 
-def app
-  ##
-  # You can handle all padrino applications using instead:
-  #   Padrino.application
-  TokyoProject.tap { |app|  }
+# You can use this method to custom specify a Rack app
+# you want rack-test to invoke:
+#
+#   app TokyoProject::App
+#   app TokyoProject::App.tap { |a| }
+#   app(TokyoProject::App) do
+#     set :foo, :bar
+#   end
+#
+def app(app = nil, &blk)
+  @app ||= block_given? ? app.instance_eval(&blk) : app
+  @app ||= Padrino.application
 end
